Add support for custom KexAlgorithms
Certain old hosts, such as SLES10-SP3, require a custom ssh
KexAlgorithms to be set, otherwise you won't be able to ssh into
these hosts from a modern system. Vagrant is stupid about this,
and so will terraform. This is a distribution specific setting,
and so, it is set by a configuration preference by selecting the
respective old distribution when using kdevops.

This adds two new tests cases, one to mimic the use case if used
by vagrant, the other to mimic the use case if used by terraform.
We do have a custom internal-for-now SLES10-SP3 vagrant image for
vagrant, however I am not sure if we'll ever have anything on the
cloud for SLES10-SP3, but hey, at least this is supported now.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>