kdevops_nodes_split_end.yaml.in: fix ansible extra_args.yml path
The ansible/ directory was used on oscheck, but once we merged
oscheck into kdevops and renamed it as its own playbook as "fstests",
we had changed the directory where we stash playbooks as "playbooks".

This is configurable, so just use the variable we set.

We also technically don't really need this as all kdevops playbooks
go out of their way to look for this file. However, we can remove
that later if we want.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>