kdevops: run update_ssh_config_vagrant role locally
SLES10-SP3 requires a custom KexAlgorithms:

KexAlgorithms diffie-hellman-group-exchange-sha1,diffie-hellman-group14-sha1,diffie-hellman-group1-sha1

Without this you won't be able to ssh into these hosts, even
though vagrant is able to ssh in, and even if we extract the
$(vagrant ssh-config) and paste it into our configuration.

We must therefore add support use a custom KexAlgorithms.

We don't yet add support for this on kdevops, we simply update
the ansible roles we have which we use, given that the change
to support older hosts such as SLES10-SP3 *also* means we
must be able to *not* use ansible for them. We currently
provision two playbooks when using vagrant:

 * devconfig
 * update_ssh_config_vagrant

We therefore remove running the playbook update_ssh_config_vagrant
from kdevops, and instead run it locally on the controller which
obviously does have support for ansible.

SLES10-SP3 has Python 2.4, and so cannot be used with ansible.
But since all the work behind the python script can be run
locally, there is absolutely no need to use ansible on the
target host.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>