kdevops: implement splitting devconfig semantics
The devconfig role does a lot by default:

  * adds your .gitconfig, .bashrc, etc
  * it refreshes your repositories (registeres system too)
  * updates your system
  * install kdev tools

If we want to test kdevops, this is a lot of work. We don't just
want to disable devconfig for testing though, we want to be able
to split this work up and pick and select what operations we want
to do. For example, for basic fast test, all we need is to do the
basic devonfig tasks asuch as copying your .gitconfig, but nothing
time consuming.

We didn't have semantics for opting out of the last items until
devconfig v3.4.9, which we uprade to. We now then allow the user
to specify if they want to run the later tasks which take time.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>