kdevops: move ssh update preference from terraform to kdevops menu
You can currently only disable updating your ssh configuration file
updating on terraform. Now that we've split out updating your ssh
configuration file from the devconfig ansible role to just us
doing it manually for you on kdevops on the Makefile target,
we can easily process the update preferences in a similar way
for vagrant.

So make ssh update configuration options generic and add support
on the vagrant side to disable the update as well. Since this
is a fast task we leave this enabled by default.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>