kdevops: unify ssh config update strategy
Terraform used to do two operations when updating the
ssh config, one for removal and another for adding the hosts,
while also creating two backup files.  This was due to paranoia,
and as there were no solid tests in place for the python script.

The Python script however now has its own set of tests, and we've
enabled travis for Continous Integration. Confidence in its use,
and confidence in avoiding regressions is pretty high now:

https://github.com/mcgrof/update_ssh_config

Both the terraform module and the vagrant ansible plugin re-use
this code as a git subtree.

So unify the strategy, and document the backup strategy now
in a simplified way.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>