kdevops: remove ansible galaxy dependency
Using ansible galaxy roles made sense prior to kconfig
integration, when we thought we'd have to fork kdevops
for any project. By using kconfig we now know we can
integrate workflow support for different kernel testing
infrastructures using the same kdevops tree and by just
having the test infrastructure be used as a kdevops
workflow.

Supporting the use of ansible galaxy roles then becomes
pointless unless we have external projects using the
ansible roles we have created. I don't see that happening
at this point. And the overhead of trying to pull in
external dependencies, and updating tons of different git
treese just to make a new kdevops release is just not
worth it.

This simplifies developmnent for kdevops. We are missing one
step, adding the update_ssh_config subtree which we will do
next. The terraform module does remain independent, but that
is fair, given others can easily take advatage of that. Its
the only area where we will be sharing code. So when we update
the code on the upstream update_ssh_config tree, we will then
have to update both the terraform module and the local ansible
update_ssh_config_vagrant role.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>