kdevops: add support for custom kdevops git tree and targets
Enterprise Linux distributions may opt to embrace kdevops as a git
subtree. In such cases then kdevops becomes a directory in their main
git tree.

We need a way to allow enterprise distros to override the git URL,
and also where the real kdevops directory is. This adds support for
that. SLES will be the first distro that uses this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>