devconfig: reboot after doing a full system upgrade
Some distributions will opt-in to remove old modules after doing
a full distribution upgrade. If you do this, and then say, try to
mount a filesystem for which you did not have the module loaded for
the system will fail telling you that it it doesn't know anything
about that filesystem. The error message is correct, but it is not
obvious.

To avoid such situations just ensure we reboot after we do a full
upgrade, if that option is enabled -- to do the upgrade.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>