vagrant: run ssh role prior to devconfig
The devconfig role does quite a bit of work, and so it can take a long
time and it may fail for may reasons. When things go bad though
we want to be able to at least log in to our systems though, so we
should instead first run the ssh update role prior to the devonfig role.

Then, since the devconfig role can fail for many things, let's just
split out running the devconfig role from vagrant provisioning. This
way our vagrant use is reduced to only bringing the initial system
up.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>