devconfig: fix label for when devconfig_repos_addon is empty
You cannot use an item from a list which will be empty if
you have disabled what adds elements into that this. In this
case if you have enabled SLE registration and disabled the
new addon repository method, ansible complains about the label.

Fix that by not relying on the items on the label.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>