fstests: fix oscheck's dependency checker and run dependency checker first
Within check_deps() the RET variable is used to check for dependencies,
however the last call in check_deps() is for check_services() and that
overwrites this local variable. Instead use a more unique variable for
check_deps().

Prior to this change we were never bailing out when checking for
dependencies, and so only check_services() was actually forcing
us to bail out.

While re-thinking this through, we want to check for deps and
then bail out if an error always occurs. We want to restrict
oscheck.sh from running tests if we have failed to meet
dependencies, but we *always* want to check for dependencies
and bail if an error was found.

This ensures that we can *always* have a dependency checker
prior to running tests.

This also allows us to also run a fist pass with ansible on the
depenency checker alone, and know that this is all we are doing
first.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>