fstests: bring forward most sanity checks on --check-deps
Prior to running check.sh there are a series of sanity checks we run,
even *after* the current --check-deps point. When doing automation
though we want sanity checks to run first, prior to a long full
run of tests. This way we *know* if a failure was a sanity deps
check or a real test case failure.

Move these late tests over so that they are captured inside the
call for --check-deps.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>