fstests: fix TEST_DEV sanity check to use -e instead of -f
The bash -f will check if the target file is a file, but will fail
if it is a block device. Although technically this can be seen
as a regression within bash, we want to adopt a use case which will
work for older versions of bash. So using -b may not be ideal.

Instead just use -e for exists.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>