fstests: move functions higher up and move variable setters up as well
There are quite a bit of functions declared after we run checks.
This makes reading the flow of oscheck difficult. We also have
some variable setters, such as when we queue section names.
We can safely move these variable setters to be called earlier,
and then just have the functional logic of checks called followed
by the running of the final check command.

This just moves code around. It should in theory not regress anything
other than the possible varible setters.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>