fstests: address missing runtime on gen_results_summary.py
gen_results_summary.py can run into a situation where no runtime
is found and so when parsing None object is used and fails to
run the script. This can happen if you've tried to run a test,
crashed, the system is not responsive and tried to somehow parse
the xunit file.

Special-case this to use 0 run time, instead of failing. We can
figure out a better value later, but failing to run when collecting
results is just not acceptable.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>