fstests: fix calling augmenting expunge list script
The script which augments our expunge list was not being run given
we recently added a recent "when" clause which depends on the
execution path run on target hosts, not on the *localhost*, and
the augmentations task runs in the namespace of localhost, not
the target host. This *should* explain why even though the when
clause is true it never runs. Instead of looking and verifying if
this is correct, and how to interpolate namespace varibles registered
locally, simply augment the script every time. We can rely on tags
for doing the right intended thing.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>