update_ssh_config.py: reorder adding hosts to the top
mwilck reported through kdevops github issue #3 [0] that
the way we should add hosts is to add them to the top of the
ssh configurationf file, since the first obtained value for each
parameter is used, more host-specific declarations should be given near
the beginning of the file, and general defaults at the end.
This is documented on the ssh_config(5) manpage:

  Since the first obtained value for each parameter is used, more
  host-specific declarations should be given near the beginning of the
  file, and general defaults at the end.

Fix this by adhering to the what ssh_config(5) recommends.

[0] https://github.com/mcgrof/kdevops/issues/3

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>