update_ssh_config.py: add IP address on Host line
On issue#3 [0] Martin Wilck suggested adding the IP address as well on
the Host line.  This seems to have been supported for years now. This
was relatively easy to support.

The logic behind this is that when you ssh with the IP address the
same configuration options will be used as well.

[0] https://github.com/mcgrof/kdevops/issues/3

Suggested-by: Martin Wilck
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>