fstests: add shortcut output expunge dir for sles
If you are going to automate testing for *any* kernel on a specific
release, you want to augment your *generic* baseline, but only if
you have no baseline already for your kernel. This strategy allows
you to override the generic baseline you have, with a kernel release
specific baseline if it is present.

This has a few implications as well. We must therefore define a shortcut
method per distribution, this cannot be generic, as far as I can tell.
Also, I'm observing that very often we have new failures, but if you
don't already have a created file, we won't know the diff unless
you have an empty file. For this later reason I will recommend moving
forward to have empty files for baselines for sections which have no
issues. This way an augment on the expunge list is exhaustive, and git
diff will truly show the delta when automated.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>