fstests: add support to print a special message when done
If you are monitoring dmesg for progress of each fstests,
you can see how many seconds a test is taking to run, but
this works up to the point we reach the last test as
fstests today has no clear way to indicate that all tests
are done.

Provide a special print message, which a knob which allows
us to later disable this, in case this gets upstream onto
fstests.

This will allow fstest test watchdogs to monitor if the last
test completed and when.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>