fstests: copy check.time files prior to run and after run
The host that controller is the system where we run ansible on,
and although the check.time might be useful on each host, it
is also valuable on the controller prior to a run, and perhaps even
after a run.

Its useful during a run so that if a controller wants to implement
an fstests watchdog it can do so without querying each host for
information for prior runs. We need to copy this data after we
wipe results directory from the prior run. We avoid implementing
an excemption to removal given that it is faster to just copy
the file after, than implement an exlcusion list.

We copy the check.time files *after* as well, so that if we want
to do some sort of post-processing we can do that as well.

Note that this information is ephemeral, however *iff* the
kernel version matches, and the section matches, it is pretty
useful information. And so we keep this only right prior to
a run.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>