kdevops: add initial kernel-ci support
This adds some initial kernel continous integration support using the
fstests workflow as an example. The idea behind this is after you
establish your baseline for a test you then want to ensure you don't
regress. To check for regressions you run a workflow test in a loop
until a failure is found.

This adds support to allow users to email reports for failures.

To run the fstests kernel-ci loop:

  make fstests-baseline-loop

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>