fstests: add an initial watchdog script
Sometimes you might end up hosing a system with a test, and
unless you have patience you may not know for a while that the
system crashed. This implements an fstests specific watchdog
which keeps track of running fstests on different hosts on
kdevops and outputs the currently running test, and expected
time left for each test.

If a test is suspected to have created an issue we annotate this.
Later on this can be used to do things like reset target suspect
crashes.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>