libvirt_user: verify apparmor is disabled
An ofter error folks can run into is that the user cannot spawn a guest,
even though all the permissions are set. The issue is caused by
apparmor.

The easy way to disable apparmor is to append apparmor=0 to your kernel
command line parameters on the grub configuration file, however there is
no easy programatic way to adjust this file for all users, and so just
complain about it. We are sure users can at least do that.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>