shared_software: disable installing postfix and bsd-mailx on debian
Installing postfix on debian-testing with ansible doesn't work as expected
as of today, January 9, 2021, we run into a chicken and an egg problem
if we also want to install bsd-mailx as postfix won't be configured if
we install postfix without a configuration.

We don't yet have kconfig logic which lets us process the postfix
configuration file prior to installing postfix, and so just skip postfix
and the watchdog software if on debian.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>