fstests: add initial btrfs support
More work is needed, but I ran into the snag that part of 'make fstests'
is to git clone kdevops on each target hosts, and later also run:

	make install

On each guest. The issue is that this target requires the filesystem
configuration upstream already, creating a chicken and egg problem.
For now just push these initial changes out so we can chug on, and
we can re-evaluate how to make adding new filesystems easier later.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>