devconfig: ensure /etc/hostname is always set
When testing with Debian testing today I noticed /etc/hostname
was not set. But I noticed it was not set when running the fstests
test, ie, make fstests-baseline. The issue was that the configuration
file was not found, ie, fstests was looking for
/var/lib/xfstests/config/buster.config but obviously the real issue was
that the hostname was not actually set properly.

So let's be paranoid and always ensure the file is set.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>