btrfs: fix section names to match host generation expecation
When using btrfs the ansible hosts file is generated using two scripts:

make V=1 hosts
.//scripts/workflows/fstests/btrfs/gen_manual_hosts.sh
.//scripts/gen_hosts.sh

The first loops around the BTRFS_SECTIONS variable, and there
we use short names instead of long names. So match the kconfig
entries to the short names. The shortnames are preferred as
its what will be used on the target hosts.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>