fstests: add support to deal with hung tests with our own fstests TestDog (TM)
When you run fstests you may end up in a situation where a tests
hung a system being tested. fstests was designed to run on the
target host, not to be a dispatcher. When working with a dispatcher,
it is up to you to detect a hung system and deal with that.

There are many ways to do this, however instead of relying on say
the watchdog software to use the kernel watchdog support, we can
easily implement our own test specific fstests watchdog (fstests
TestDog (TM)).

Since we are dealing with virtual guests for most tests, if we cannot ssh
into the system, then we can safely assume things went south. Likewise,
all we need then is a point to inform the fstests TestDog (TM) that
monitoring for tests is OK, at that point we can let the TestDog ssh
into the system and assess the health of the system. For now, all
we care about is detecting we can scrap the kernel log, and get the
last known fstests test which we ran. If we cannot get this information,
or we cannot log in, we know for sure we likely have a hung system.

A hang is the worst time of issue we can run into.  If you are working
on a baseline, you want to be immediately notified of a hang, and so
we add support to deal with the hang by immediately trying to quiesce
all tests, ansible, and just email admins about the situation.

Unfortunately, ansible doesn't make it easy for us to quiesce it
from outside of ansible. Sending a SIGALRM can be used to SIGALRM
ansible processes, but not its own children. So it would seem we
need to colb for ansible child processes and then SIGTERM them,
prior to sendign SIGALRM to the ansible processes. That also doesn't
yet address your own child processes which called ansible. This
quiesce effort is simply work in progress for now.

The easier thing to do, is simply to use ansible as the dispatcher,
and deal with this issue within ansible. This means, first collecting
the list of tests you want to run. And then run each test, one by one.
This allows us to give ansible a notion of a time limit for each test.
If the timelimit passes a threshold we known its a hang, whether or
not its an ssh / network issue or a real hang.

But this means implementing the dispatcher for ansible. This is ideal
though, given we want to move to the ability to then allow the
dispatcher to say run each test within a container. That would allow
us to parallelize all tests.

That's next.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>