lib/fstests: add guard for validation of watchdog results
A status update on fstests using the kernel's ring buffer
is only valid so long as the fstests check process is
running.

We have to be careful with this heuristic, since, a timeout
also is reflective of a possible hang. This does mean
introducing two more ssh commands to each target on the
watchdog, however this shouldn't be an issue. It is better
to affirm that fstests is running than to report a hang on
a false possitive.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>