kernel-ci: add steady state goal support
If you enable kernel-ci, you currently just run the loop until a failure
happens. This is pessimistic, and doesn't give us a goal to achieve. If
we never fail, we'd run forever.

Add support to state *when* we have succeeded, we borrow the term used
for disk IO "steady-state" goal on modern drives for this, and refer
to this as "a test steady-state goal".

We accept that some tests may wish to lower or increase their own
steady state goal, and so we refer to the common goal as generic.
It is up to each workflow to adopt/ignore this value. Distros can
easily override this later if they so wish.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>