bootlinux: simplify tag / version usage
Right now we use a tag version, and then add a "v" prefix, but this
doesn't work well for all trees. In fact, this is just silly. The
right thing to do is to clarify the user specifies the tag and we
just use that.

This just means changing the template file names to not use the prefix.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>