kernel-ci: add support for custom email subject prefixes
A distribution may choose to also track linux stable or linux-next,
or Linus' vanilla tree. In such cases, where we build those kernels
and boot into them, we want the kernel-ci email report to allow us
to distinguish the goal and the fact that we are on a custom kernel.

This adds support for this.

Likewise, it also allows us to customize the kernel-ci email report
based on the target distribution, if you are not on a custom kernel.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>