workflows: add support for none and combined workflows
Since we have added support for a custom kernel-ci topic
to be used on the subject for emails, we *assume* that
CONFIG_BOOTLINUX is not always enabled, but that's not
true.

And so, if you end up using a distro for your vanilla
distro kernel you'd end up getting a linux custom build
prefix, even though you are just using a distro kernel.

To fix this we just separate make a default workflow to
be do nothing, and then also split up the fstests workflow
to to be detached from the BOOTLINUX workflow. We do this
later part by adding a new workflow specification which
clarifies you might also want to do both things:

  * boot into a custom kernel
  * run fstests

This way, if you are just running fstests, but are on a
distro kernel, we now have the semantics for determining
this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>