fstests_watchdog: detect login failures as hangs
If we can't log in anymore, experience shows the first issue
the watchdog will pick up is a "Uname-issue", this is triggered
when lib/kssh.py uname function tried to issue an ssh command
and tried to run "uname" but the ssh command failed right away
with an error code.

We can rebrand this error to any other label later, but for
now all we care for is to detect the failure, and report it
as a possible hang.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>