devconfig: use vars tag for setting variables
We used an snmp tag just so we can be lazy and install snmp easily
by just specifying the snmp tag to the playbook. But we want to do
the same thing for other packages, such as with kdump later, so
just use one tag for variable setting, that way admins who want
to install a speecific package and just add two tags to install it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>