devconfig: add initial kdump support
We only enable this on sles15 so far, as that's where I'm testing
this. As other distros get support we can expand on this and generalize.
Also notice that the update-grub stuff is already implemented
generically for other distros on the bootlinux ansible role, so we
can later just share the code with that ansible role.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>