fstests: add support to generate check.time distribution files
If you have a confidence in your baseline with for example 100 loops and
it takes you 5 days to run those 100 loops, you may end up realizing
that you somehow want to create even *more* confidence by running faster
tests in a loop on spare systems. You'd have to ask:

   which are the tests which run the fastest?

The check.time.distribution file shows you this. It also show you
how much percentage of coverage that represents in consideration for
all the number of tests which actually ran. You can then make decisions
and trade offs on running tests in spare systems in a loop elsewhere
based on this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>