fstests: fix section lookup and add section verifier
When we use ansible_hostname and /etc/hosts is populated as follows:

127.0.1.1 linux-v5.10.y-btrfs-raid56 linux-v5

ansible_hostname will map to while linux-v5
ansible_ssh_host will map to linux-v5.10.y-btrfs-raid56.

So we need to use ansible_ssh_host for processing the section
name.

While at it, add an earlier verifier for the section name.
This way we can bail out early in case its not there.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>