fstests: fix augment process bug
We augment the expunge list when *.bad files are found, however
sometimes after this step files were observed to be empty or
not being stripped somehow of content. I've traced down to
not being a bug with the python expunge script but instead
with the sort post processing script used. It would seem that
using tee as a hack is not as reliable as one would hope, specially
if one is appending data to the file sequentially, one after the
other to the same exact file.

A curious thing was that this issue is observed on btrfs, xfs, and
even nfs, but not on ext4. Fix the hack so to make things clear.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>