fstests: split up finding the files and processing them
When looking into the odd bug I was seeing on processsing the
expunge files I initially could not tell what was causing the
issue. To be even safer, let's split off the os walk from
post processing of the files.

This also lets us move the processing of the files with one less
indent and makes the code easier to read.

While at it, add a verbose argument to allow us to inspect things
step by step.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>