fstests: use /var/lib/xfstests/configs.config over /var/lib/xfstests/configs/$HOST.config
I've seen that depending on the distribution you may end up with
a short name version of the host on /etc/hosts, and then this may
in turn be used for a lot of assumptions. Instead of relying
on the HOST variable, just read the host independent file,
/var/lib/xfstests/configs/.config which we always install now
as well.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>