fstests: all a valid section to include the name defaults
For ext4 we are going to name the default section ext4_defaults,
however oscheck.sh intended to grep out the [default] section so
we don't count it. The grep out however was also removing section
names like [ext4_defaults] which we want to use for ext4, so fix
this by greping out the [defaults] section explicitly first.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>