fstests: detect early bail on steady state goal
I have detected a situation where we bail the kernel-ci loop,
detect no failures, and yet the loop count does not match the
steady state goal. This happened when no failures were found,
and no OOMs on the hypervisor, or no failures on the hypervisor
detected.

We'll see if this creeps up again.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>