workflows: select KDEVOPS_ANSIBLE_PROVISION_ENABLE when enabling workflows
If you enable a workflow but disable ansible configuration you'll end up
with commands like 'make blktests' failing with obscure messages like:

	you probably need to set the interpreter. See stdout/stderr for the
	exact error. The following modules failed to execute: setup

This is because the python interpreter ends up empty on the hosts file.
Fix this by forcing the proper generation for of the python interpreter
always when we enable a workflow.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>