install_vagrant_boxes: add support for json formats on urls
If we are using a json file for a box URL we must stick to
the name given to the box upstream. So respect this. Likewise
we also need to be explicit about the provider used.

This cakes in libvirt support. We can expand to support other
types of boxes when someone find the time to add such boxes.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>