blktests: add support to generate results directory from last-run
blktests works slightly different than fstests. Other than *.bad files,
it also carries with it *.dmesg files *and*... wait for it... the empty
test file name also has metadata about the test, test description and
status, as well as runtime.

The status file is actually what we use to extract the last run time,
and is already used by the watchdog to estimate completion time for a
new run, provided we had a last-run recorded.

The *.bad and *.dmesg files, along with the status files are *prime*
targets for us to use for automation for filing bugs, and so we want
to keep them around.

We add a new python script which scrapes the last-run directory, where
we copy all result files, it then copies all of that directory to a new
results directory for the target kernel, and then removes any files for
tests which succeeded. What is left in place then is only meta data and
actual *useful* information for any failure that may have occurred.

The way the kdevops blktests oscheck.sh script works, is slightly different than
the         kdevops fstests  oscheck.sh script work, in that the expunge
list for blktests is inferred by looking for all *.bad and *.dmesg files.
The expunge list file is simply a convenience for blktests, while for fstests
it it is the *source* for expunges. That is, in fstests, oscheck.sh
reads expunge files and appends to fstests's check arguments a series of
-E arguments with these expunges files. For blktests, we also generate
the expunge list dynamically on the fly based on present prior results
from a prior run for a target kernel, by looking for *.bad and *.dmesg
files. The expunge list then for blktests is just a convenience to let
us append metadata such as bugzilla information and developer
interpretations of failures.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>