blktests: add support to also look at the manual expunge list
If you have a blktests which crashes  the system, such as nbd/003 on
Debian testing, you will *not* end up with a respective test *.dmesg or
*.bad file, given blktest's check does not look for hangs. kdevops'
watchdog however picks up on these hangs, and reports them.

And so we need to track manually added expunge entries on the single
expunge list in case we have those type of failures. This also allows
us to be lazy and not add *.dmesg or *.bad files for certain failures,
when perhaps our tools failed to add them for us. A tooling issue
or typo seems to have happned to scsi/004 at least for Debian testing.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>