kotd: complete kernel continous integration support with a new loop command
The ideal situation is you want a loop which upgrades a target series of
hosts to the latest supported kernel, and then runs a target test, only
to fail and and report back when an issue is found.

Since we already have support for a kernel-ci loop for testing fstests
and blktests, and we now have a new kotd command, all that is missing is
a script which brings these two together.

A series of .kotd.* files are used to track this work. This now
completes full kernel continous integration support.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>