blktests: fix target kdevops dir when generating symlink
A parent project may fork off kdevops and use kdevops only as a subdir.
Such is the case in how SUSE uses kdevops, it uses kdevops as a git
subtree, and so kdevops is just a directory under sle-kdevops/, ie
sle-kdevops/kdevops. To account for this use case the {{ kdevops_dir }}
variable was provided a while ago. Use this variable to fix installing
symlinks for kdevops on SLES releases. This fixes the installation of
the only symlink for kdevops blktests purposes, the oscheck.sh script.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>