blktests: fix installation of dependencies for SUSE releases
There is no packages for nbd-client or nbd-server on SLE, and so
we must compile it from source. There is also no docbook2man which
is required for ./autogen.sh to complete when generating documentation,
so instead of trying to muck with repos to add that dependency just
touch the man pages as empty files, we don't care for them as we're
testing things as part of a kernel-ci system, not a kernel-ci readme
database :). Developers will use their own systems where they can
happily read all the docs they need.

This allows us to run this on SLE:

  o make blktests

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>