blktests: finish renaming stale variables from fstests work
A lot of the kdevos blktests stuff was heavily inspired from kdevops
fstests work, and so we inhereted tons of defaults from fstests. The
defaults are set by kconfig though, and so we don't run the risk of
not having these varaible set. However we must rename these and
clear the clutter so to avoid any confusion.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>