kill_pids.sh: fix when using in the watchdog
When we added support to bailing out of the kernel-ci loop right
away when we are running kill_pids.sh manually we failed to address
the fact that we *want* to proceed with post processing of the results
when processes are killed from the watchdog.

So add a flag to ensure we don't make an exception when we are killing
the PIDs so we continue to get get email reports on failures caught by
the watchdog.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>