kdevops: add a new reboot-limit test workflow
If you run tests which reboot systems you may find that sometimes you
end up with some odd false positives which make no sense. The issue
may just be issues in the reboot process, whether the network does
not get set up properly, or whatever. This is specially true if you
are doing this setup in a virtualized environment. Consider things
like the DHCP server. There are many points of failure.

The reboot-limit workflow is designed to test how many reboots can
be completed successfully without issue. While at it, it also adds
the optiont to collect a few statistics for reboots, such as results
from systemd-analyze, and also gives the opportunity to purposely
crash a system after x number of reboots. The crash lets you test
things like custom watchdogs you might have implemented. For now
we enable forcing the watchdog as configured per systemd, and
since we always install kdump, it should mean that custom watchdogs
can be extended to scape for crashes on /var/crash on target hosts.
We leave this later effort for future excercise.

Since we don't have much documentation for how to add a new simple
workflow we take the time to ensure this ones serves that purpose.
Eventually we should move some of this documentation into another
format, but for now we can just point folks to this workflow to get
an idea of how to add more.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>