devconfig: replace consoleblank entry
If a system has consoleblank entry we want to replace that line
completely, and assume a default install was done. This was
noticed on Debian, and our console grep assumed incorrectly that
these systems had a console setup.

Fix that by dealing with consoleblank entries on its own and
fully replacing the command line entry when that is present.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>