kconfigs/workflows/Kconfig: allow both blktests and fstests to be enabled
Some folks may want to enable both fstests and blktests.
Allow for this. Before this patch you were only allowed to
pick one. Now you have to specify that you really want to
not make this a dedicated test setup but rather you are
going to want to pick what you want to test.

If you enable to test both fstests and blktests you will not
get the host file generated in case you enabled virtualization
or cloud setup for the target test given we'd have two tests
enabled. What hosts do we enable? In the future we might opt-in
to allow to augment the number of nodes to be commulative, but
for now we simplify and just disable that option, and go with
the default kdevops nodes configuration.

In practice this means that if you don't enable a dedicated
target test you will end up with two hosts if you enabled
virtualization, kdevops, and kdevops-dev. I find myself needing
to test both fs and block often enough this is now useful.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>